+++ .gitignore
@@ -102,3 +102,4 @@ ENV/
 
 # paper
 log.txt
+.idea/
