+++ paper/librarian.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 import os
-import re
 import requests
 import yaml
 from pyquery import PyQuery
@@ -45,24 +44,6 @@ class Librarian:
             paper = extract_paper_from(pq_div)
             papers.append(paper)
         return papers
-
-    def get_user_input(self, papers):
-        user_input = input('Paper to download [0-9] or all: ')
-        while not self._is_valid_input(user_input, papers):
-            user_input = input('Paper to download [0-9] or all: ')
-        return user_input
-
-    def _is_valid_input(self, user_input, papers):
-        if user_input == 'all':
-            return True
-        elif re.match(r'[0-9]', user_input) is None:
-            return False
-        elif papers[int(user_input)]['url'] is None:
-            return False
-        elif papers[int(user_input)]['authors'] == []:
-            return False
-        else:
-            return True
 
     def select(self, papers, user_input):
         if user_input == 'all':
